+++ changelog.md
@@ -24,6 +24,8 @@ Codion Change Log
 - AttributeDefinition.maximumFractionDigits() and decimalRoundingMode() renamed to fractionDigits() and roundingMode().
 - AttributeDefinition.Builder.numberFormatGrouping() renamed numberGrouping().
 - AttributeDefinition.Builder.localeDateTimePattern() renamed dateTimePattern().
+### is.codion.framework.model
+- EntityEditModel.EditEvent removed.
 ### is.codion.plugin.flatlaf.intellij
 - Catppuccin themes added.
 - VSCode themes added.